requirements.txt
# requirements.txt
#torch>=2.0.0
#torchvision>=0.15.0
#numpy>=1.24.0
#matplotlib>=3.7.0
#pillow>=9.4.0
#tqdm>=4.64.0
#wandb>=0.15.0
#opencv-python>=4.7.0
#scipy>=1.10.0
#scikit-image>=0.20.0
#transformers>=4.30.0
#diffusers>=0.17.0
#ftfy>=6.1.0
#accelerate>=0.20.0
#tokenizers>=0.13.0
#clip>=0.2.0

torch>=1.9.0
torchvision>=0.10.0
diffusers>=0.14.0
transformers>=4.25.1
accelerate>=0.16.0
huggingface-hub>=0.13.0
numpy>=1.20.0
matplotlib>=3.4.0
tqdm>=4.62.0
Pillow>=9.0.0
scipy>=1.7.0
scikit-image>=0.18.0
ftfy>=6.0.0
opencv-python>=4.5.0
wandb>=0.13.0
clip-anytorch>=2.0.0
lpips>=0.1.4
tokenizers>=0.13.0